requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-chroma
langchain-huggingface
sentence-transformers
chromadb
python-dotenv
